pyproject.toml
[tool.poetry]
name = "ylab"
version = "1"
description = ""
authors = ["Yryskeldi"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
Django = "3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
